edit not working yet

+++ src/main/java/logViewer/Model/FileFormat.java
@@ -9,8 +9,7 @@ public class FileFormat {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
+  private int id;
   private String name;
   // File extension
   private String fileType;
@@ -21,11 +20,11 @@ public class FileFormat {
   public FileFormat() {
   }
 
-  public Long getId() {
+  public int getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(int id) {
     this.id = id;
   }
